+++ src/model/ComparePhoneContact.java
@@ -6,7 +6,7 @@ public class ComparePhoneContact implements Comparator<Contact>{
 
 	
 	public int compare(Contact c1, Contact c2) {
-		return c1.getPhone()- c2.getPhone();
+		return (int) (c1.getPhone()- c2.getPhone());	
 	}
 
 }
